Move LocalResult tests into results file and add tests for absolute path exists call
