+++ tests/engine/results/test_results.py
@@ -1,10 +1,19 @@
+import os
 import json
+import tempfile
 from typing import Union
 
+import cloudpickle
 import pytest
 
 import prefect
-from prefect.engine.results import ConstantResult, PrefectResult, SecretResult
+from prefect import config
+from prefect.engine.results import (
+    ConstantResult,
+    LocalResult,
+    PrefectResult,
+    SecretResult,
+)
 from prefect.tasks.core.constants import Constant
 from prefect.tasks.secrets import PrefectSecret
 
@@ -120,3 +129,65 @@ class TestPrefectResult:
         result = PrefectResult()
         assert result.exists(json.dumps(value)) is True
         assert result.exists(value) is False
+
+
+class TestLocalResult:
+    @pytest.fixture(scope="class")
+    def tmp_dir(self):
+        with tempfile.TemporaryDirectory() as tmp:
+            yield tmp
+
+    def test_local_result_initializes_with_no_args(self):
+        result = LocalResult()
+        assert result.dir == os.path.join(config.home_dir, "results")
+        assert result.value == None
+
+    def test_local_result_initializes_with_dir(self):
+        root_dir = os.path.abspath(os.sep)
+        result = LocalResult(dir=root_dir)
+        assert result.dir == root_dir
+
+    def test_local_result_writes_using_rendered_template_name(self, tmp_dir):
+        result = LocalResult(dir=tmp_dir, location="{thing}.txt")
+        new_result = result.write("so-much-data", thing=42)
+        assert new_result.location == "42.txt"
+        assert new_result.value == "so-much-data"
+
+    def test_local_result_cleverly_redirects_prefect_defaults(self):
+        result = LocalResult(dir=config.home_dir)
+        assert result.dir == os.path.join(config.home_dir, "results")
+
+    @pytest.mark.parametrize("res", [42, "stringy", None, type(None)])
+    def test_local_result_writes_to_dir(self, tmp_dir, res):
+        result = LocalResult(dir=tmp_dir, location="test.txt")
+        fpath = result.write(res).location
+        assert isinstance(fpath, str)
+        assert fpath == "test.txt"
+
+        with open(os.path.join(tmp_dir, fpath), "rb") as f:
+            val = f.read()
+        assert isinstance(val, bytes)
+
+    @pytest.mark.parametrize("res", [42, "stringy", None, type(None)])
+    def test_local_result_writes_and_reads(self, tmp_dir, res):
+        result = LocalResult(dir=tmp_dir, location="test.txt")
+        final = result.read(result.write(res).location)
+        assert final.value == res
+
+    def test_local_result_is_pickleable(self):
+        result = LocalResult(dir="root")
+        new = cloudpickle.loads(cloudpickle.dumps(result))
+        assert isinstance(new, LocalResult)
+
+    def test_local_result_writes_and_exists(self, tmp_dir):
+        result = LocalResult(dir=tmp_dir, location="{thing}.txt")
+        assert result.exists("43.txt") is False
+        new_result = result.write("so-much-data", thing=43)
+        assert result.exists("43.txt") is True
+
+    def test_local_exists_full_path(self, tmp_dir):
+        result = LocalResult(dir=tmp_dir, location="{thing}.txt")
+        assert result.exists("44.txt") is False
+        new_result = result.write("so-much-data", thing=44)
+        assert result.exists("44.txt") is True
+        assert result.exists(os.path.join(tmp_dir, "44.txt")) is True
